ISE-289
Added version numbers to the dependancy declaration in the metadata
files of each role cookbook.

ex: "depend 'role-base-nativex', '0.1.0'

This explicitly requires the noted version number of the running
cookbook be used. Then in the chef environment we specify the version of
the "role cookbook". Allowing for fine grain managment and control on the
cookbook versions used in any given environment.